Make stream include usage as optional
